moved train, generate, and predict into one file
